Version V1.0.2  1.add serialport and socket auto reconnect 2.fix some bugs

+++ sdk/src/impl/include/serial/nvilidar_serial_list_unix.h
@@ -71,7 +71,7 @@ glob(const vector<string>& patterns)
         glob_retval = glob(iter->c_str(), GLOB_APPEND, NULL, &glob_results);
     }
 
-    for(size_t path_index = 0; path_index < glob_results.gl_pathc; path_index++)
+    for(int path_index = 0; path_index < glob_results.gl_pathc; path_index++)
     {
         paths_found.push_back(glob_results.gl_pathv[path_index]);
     }
